Feat(SideMain.js): Sidebar와 Main으로 구성된 레이아웃 생성
기존의 LearningPage의 레이아웃을 별도로 SideMain으로 재구성
포럼에도 사용하기 위해서 재구성함
prop.children 사용하여 sidebar 구현
Outlet 사용하여 Main 내용 구현
